:recycle: refactor(#168/user): user 도메인 관련 클래스 ninedot-domain으로 이동
