+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/user/exception/UserException.java
@@ -1,0 +1,11 @@
+package org.sopt36.ninedotserver.user.exception;
+
+
+import org.sopt36.ninedotserver.exception.BusinessException;
+
+public class UserException extends BusinessException {
+
+    public UserException(UserErrorCode errorCode) {
+        super(errorCode);
+    }
+}
