--- src/main/java/org/sopt36/ninedotserver/user/exception/UserException.java
@@ -1,11 +1,0 @@
-package org.sopt36.ninedotserver.user.exception;
-
-import org.sopt36.ninedotserver.global.exception.BusinessException;
-import org.sopt36.ninedotserver.global.exception.ErrorCode;
-
-public class UserException extends BusinessException {
-
-    public UserException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
